Split renderer
